requirements.txt
openai
requests==2.25.1
requests-oauthlib==1.3.1
urllib3==1.26.5
wadllib==1.3.6
whichcraft==0.6.1
yarl==1.9.2
zipp==1.0.0
pyuwsgi
Flask
Flask-Cors
GitPython
boto3